requirements.txt
black==22.6.0 
isort==5.10.1
jupyterlab==4.0.12
jupyterlab-code-formatter==2.2.1
numpy==1.26.2
onnx==1.15.0
onnxruntime==1.16.2
opencv-python-headless==4.8.1.78
pandas==2.1.3
pyzipper==0.3.6
ruamel.yaml==0.18.5
scikit-learn==1.3.2
seaborn==0.13.2
torch==2.1.1
torcheval==0.0.7 
torchvision==0.16.1
tqdm==4.64.0